refactor type checks to own file
